fix: Eagerly initialize TurboModule because Proxies are slow
